use linear scale in quiet SEE pruning
--------------------------------------------------
Results of CherryDev vs CherryMain (8+0.08, 1t, 16MB, UHO_Lichess.epd):
Elo: 10.36 +/- 5.77, nElo: 17.85 +/- 9.94
LOS: 99.98 %, DrawRatio: 46.85 %, PairsRatio: 1.20
Games: 4696, Wins: 1218, Losses: 1078, Draws: 2400, Points: 2418.0 (51.49 %)
Ptnml(0-2): [44, 523, 1100, 611, 70], WL/DD Ratio: 0.74
LLR: 2.98 (101.2%) (-2.94, 2.94) [0.00, 5.00]
--------------------------------------------------

+++ src/cherry/search/weights.rs
@@ -129,7 +129,7 @@ weights! {
     futile_base       | FUTILE_BASE:       i32 => 93,
     futile_scale      | FUTILE_SCALE:      i32 => 79,
     see_depth         | SEE_DEPTH:         i32 => 10240,
-    see_quiet_scale   | SEE_QUIET_SCALE:   i64 => -89,
+    see_quiet_scale   | SEE_QUIET_SCALE:   i32 => -89,
     see_tactic_scale  | SEE_TACTIC_SCALE:  i32 => -62,
     cont_margin       | CONT_MARGIN:       i32 => -3268,
